+++ Source/Creation/Frame/Content/ID3RecordingYearFrameCreator.swift
@@ -10,7 +10,7 @@ import Foundation
 class ID3RecordingYearFrameCreator: ID3StringFrameCreator {
     override func createFrames(id3Tag: ID3Tag, tag: [UInt8]) -> [UInt8] {
         if id3Tag.properties.version < .version4, let year = id3Tag.recordingDateTime?.date?.year {
-            return createFrameUsing(frameType: .RecordingYear, content: year, id3Tag: id3Tag, andAddItTo: tag)
+            return createFrameUsing(frameType: .RecordingYear, content: String(year), id3Tag: id3Tag, andAddItTo: tag)
         }
         return super.createFrames(id3Tag: id3Tag, tag: tag)
     }