+++ Source/Parsing/Frame/Content/Operation/Factory/ID3RecordingYearFrameContentParsingOperationFactory.swift
@@ -10,7 +10,7 @@ import Foundation
 class ID3RecordingYearFrameContentParsingOperationFactory {
     static func make() -> ID3FrameStringContentParsingOperation {
         return ID3FrameStringContentParsingOperationFactory.make() { (id3Tag: ID3Tag, content: String) in
-            id3Tag.recordingDateTime?.date?.year = content
+            id3Tag.recordingDateTime?.date?.year = Int(content)
         }
     }
 }