+++ Tests/Acceptance/ID3TagEditorTestAcceptanceTest.swift
@@ -94,7 +94,7 @@ class ID3TagEditorTest: XCTestCase {
         XCTAssertEqual(id3Tag?.attachedPictures?[1].type, .BackCover)
         XCTAssertEqual(id3Tag?.genre?.identifier, .Metal)
         XCTAssertEqual(id3Tag?.genre?.description, "Metalcore")
-        XCTAssertEqual(id3Tag?.recordingDateTime?.date?.year, "2018")
+        XCTAssertEqual(id3Tag?.recordingDateTime?.date?.year, 2018)
         XCTAssertEqual(id3Tag?.trackPosition?.position, 2)
         XCTAssertEqual(id3Tag?.trackPosition?.totalTracks, 9)
     }
@@ -242,7 +242,7 @@ class ID3TagEditorTest: XCTestCase {
                 albumArtist: "A New Album Artist",
                 album: "A New Album",
                 title: "A New title",
-                recordingDateTime: RecordingDateTime(date: RecordingDate(day: nil, month: nil, year: "2018"), time: nil),
+                recordingDateTime: RecordingDateTime(date: RecordingDate(day: nil, month: nil, year: 2018), time: nil),
                 genre: Genre(genre: .Metal, description: "Metalcore"),
                 attachedPictures: [
                     AttachedPicture(picture: artFront, type: .FrontCover, format: .Jpeg),
@@ -271,7 +271,7 @@ class ID3TagEditorTest: XCTestCase {
         XCTAssertEqual(id3Tag?.title, "Om Tryumbacom")
         XCTAssertEqual(id3Tag?.artist, "Laraaji")
         XCTAssertEqual(id3Tag?.album, "Vision Songs Vol. 1")
-        XCTAssertEqual(id3Tag?.recordingDateTime?.date?.year, "2018")
+        XCTAssertEqual(id3Tag?.recordingDateTime?.date?.year, 2018)
         XCTAssertEqual(id3Tag?.trackPosition?.position, 10)
     }
 }