+++ Tests/Parsing/Frame/Content/Operation/ID3RecordingTimestampOperationTest.swift
@@ -17,12 +17,12 @@ class ID3RecordingTimeFrameContentParsingOperationTest: XCTestCase {
 
         recordingTimeParsingOperation.parse(frame: frameV4utf8Valid(), id3Tag: id3Tag)
         
-        XCTAssertEqual(id3Tag.recordingDateTime?.date?.day, "27")
-        XCTAssertEqual(id3Tag.recordingDateTime?.date?.month, "7")
-        XCTAssertEqual(id3Tag.recordingDateTime?.date?.year, "2018")
-        XCTAssertEqual(id3Tag.recordingDateTime?.time?.hour, "11")
-        XCTAssertEqual(id3Tag.recordingDateTime?.time?.minute, "35")
-        XCTAssertEqual(id3Tag.recordingDateTime?.time?.second, "10")
+        XCTAssertEqual(id3Tag.recordingDateTime?.date?.day, 27)
+        XCTAssertEqual(id3Tag.recordingDateTime?.date?.month, 7)
+        XCTAssertEqual(id3Tag.recordingDateTime?.date?.year, 2018)
+        XCTAssertEqual(id3Tag.recordingDateTime?.time?.hour, 11)
+        XCTAssertEqual(id3Tag.recordingDateTime?.time?.minute, 35)
+        XCTAssertEqual(id3Tag.recordingDateTime?.time?.second, 10)
     }
     
     func testframeContentWithYearInsteadOfTimestamp() {
@@ -35,7 +35,7 @@ class ID3RecordingTimeFrameContentParsingOperationTest: XCTestCase {
         
         XCTAssertNil(id3Tag.recordingDateTime?.date?.day)
         XCTAssertNil(id3Tag.recordingDateTime?.date?.month)
-        XCTAssertEqual(id3Tag.recordingDateTime?.date?.year, "2016")
+        XCTAssertEqual(id3Tag.recordingDateTime?.date?.year, 2016)
         XCTAssertNil(id3Tag.recordingDateTime?.time?.hour)
         XCTAssertNil(id3Tag.recordingDateTime?.time?.minute)
         XCTAssertNil(id3Tag.recordingDateTime?.time?.second)